use waypt_new instead of explict xcalloc - see message to -code about this. Add explict casts on allocators. gpsutil: add test for unknown alt on output. jeeps: don't use 'true' in unused (sigh) code. i65.anr.gpx, reference/netstumbler.mps: regenerate for alt change.

+++ trunk/gpsbabel/tiger.c
@@ -46,7 +46,7 @@ static char *iconismarker = NULL;
 
 int scalev;
 int short_length;
-int thresh_days;
+double thresh_days;
 
 /*
  *   The code bracketed by CLICKMAP is to generate clickable image maps
